requirements.txt

ipython
ipyparallel
anyconfig
PyYAML
sqlalchemy
wordfreq
scandir
pytest
pytest-env
invoke
factory_boy
mpi4py
clint
click
psutil
textblob
numpy
scipy
jupyter
matplotlib
sklearn
bs4
lxml
joblib
pandas
statsmodels
patsy
boltons
attrs
